+++ find-unused-exports.mjs
@@ -1,9 +1,10 @@
 #!/usr/bin/env node
 // @ts-check
 
+import { relative } from "node:path";
+
 import arg from "arg";
 import { bold, dim, green, red, underline } from "kleur/colors";
-import { relative } from "node:path";
 
 import CliError from "./CliError.mjs";
 import errorConsole from "./errorConsole.mjs";