+++ findUnusedExports.mjs
@@ -36,12 +36,16 @@ import scanModuleCode from "./scanModuleCode.mjs";
  *   [moduleFilePath: string]: import("./scanModuleCode.mjs").ModuleExports,
  * }>} Map of module file paths and unused module exports.
  */
-export default async function findUnusedExports({
-  cwd = process.cwd(),
-  moduleGlob = MODULE_GLOB,
-  resolveFileExtensions,
-  resolveIndexFiles = false,
-} = {}) {
+export default async function findUnusedExports(options = {}) {
+  // Avoid function parameter destructuring to workaround this TypeScript bug:
+  // https://github.com/microsoft/TypeScript/issues/50286
+  const {
+    cwd = process.cwd(),
+    moduleGlob = MODULE_GLOB,
+    resolveFileExtensions,
+    resolveIndexFiles = false,
+  } = options;
+
   if (typeof cwd !== "string")
     throw new TypeError("Option `cwd` must be a string.");