+++ changelog.md
@@ -12,6 +12,10 @@
 
 - Enable TypeScript syntax when parsing `.mts` and `.cts` files with Babel, even when the project has no Babel config for TypeScript.
 
+### Patch
+
+- Updated dependencies.
+
 ## 4.0.0
 
 ### Major