+++ changelog.md
@@ -1,6 +1,6 @@
 # find-unused-exports changelog
 
-## Next
+## 5.0.0
 
 ### Major